modularize structs
